refactor: change package to influxdb3
