+++ influxdb3/client_write_test.go
@@ -1,4 +1,4 @@
-package influx
+package influxdb3
 
 import (
 	"compress/gzip"
