+++ influxdb3/point.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by MIT
 // license that can be found in the LICENSE file.
 
-package influx
+package influxdb3
 
 import (
 	"fmt"
