+++ influxdb3/point_test.go
@@ -1,4 +1,4 @@
-package influx
+package influxdb3
 
 import (
 	"fmt"
