+++ influxdb3/query_iterator.go
@@ -1,4 +1,4 @@
-package influx
+package influxdb3
 
 import (
 	"fmt"
