+++ influxdb3/reflection.go
@@ -1,4 +1,4 @@
-package influx
+package influxdb3
 
 import (
 	"fmt"
